+++ test/unit_tests/test_eegneuralnet.py
@@ -19,6 +19,8 @@ from braindecode import EEGClassifier, EEGRegressor
 from braindecode.training import CroppedLoss
 from braindecode.models.base import EEGModuleMixin
 from braindecode.datasets import WindowsDataset, BaseConcatDataset
+# from braindecode.models.util import models_dict
+from braindecode.models.shallow_fbcsp import ShallowFBCSPNet
 
 
 class MockDataset(torch.utils.data.Dataset):
@@ -391,3 +393,25 @@ def test_initialized_module(eegneuralnet_cls, preds, caplog, Xy):
     assert net.module_.n_outputs == 12
     assert net.module_.n_chans == 12
     assert net.module_.n_times == 12
+
+
+# @pytest.mark.parametrize("model_name,model_cls", models_dict.items())
+def test_module_name(eegneuralnet_cls):
+    net = eegneuralnet_cls(
+        "ShallowFBCSPNet",
+        module__n_outputs=4,
+        module__n_chans=3,
+        module__n_times=100,
+        cropped=False,
+    )
+    net.initialize()
+    assert isinstance(net.module_, ShallowFBCSPNet)
+
+
+def test_unknown_module_name(eegneuralnet_cls):
+    net = eegneuralnet_cls(
+        "InexistentModel",
+    )
+    with pytest.raises(ValueError) as excinfo:
+        net.initialize()
+    assert "Unknown model name" in str(excinfo.value)